chore: Move certificate type to enums
