+++ backend/src/main/java/com/security/pki/TestCertificateSeeder.java
@@ -3,16 +3,14 @@ package com.security.pki;
 import com.security.pki.certificate.dtos.CreateRootCertificateRequest;
 import com.security.pki.certificate.models.Certificate;
 import com.security.pki.certificate.models.Issuer;
-import com.security.pki.certificate.models.Status;
+import com.security.pki.certificate.enums.Status;
 import com.security.pki.certificate.models.Subject;
 import com.security.pki.certificate.repositories.CertificateRepository;
-import com.security.pki.certificate.services.CertificateService;
 import com.security.pki.certificate.services.CryptoService;
 import com.security.pki.certificate.utils.CertificateGenerator;
 import com.security.pki.certificate.utils.CertificateUtils;
 import com.security.pki.user.models.User;
 import com.security.pki.user.repository.UserRepository;
-import com.security.pki.user.services.UserService;
 import lombok.RequiredArgsConstructor;
 import org.bouncycastle.asn1.x500.X500Name;
 import org.bouncycastle.asn1.x500.X500NameBuilder;
@@ -26,9 +24,6 @@ import javax.crypto.SecretKey;
 import java.math.BigInteger;
 import java.security.GeneralSecurityException;
 import java.security.KeyPair;
-import java.security.NoSuchAlgorithmException;
-import java.security.NoSuchProviderException;
-import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.time.LocalDate;
 import java.time.ZoneId;
