+++ backend/src/main/java/com/security/pki/certificate/dtos/CertificateResponseDto.java
@@ -1,6 +1,6 @@
 package com.security.pki.certificate.dtos;
 
-import com.security.pki.certificate.models.CertificateType;
+import com.security.pki.certificate.enums.CertificateType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
